+++ test/utils.go
@@ -22,6 +22,8 @@ import (
 	"reflect"
 	"time"
 
+	"github.com/stretchr/testify/mock"
+
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/watch"
@@ -32,13 +34,16 @@ import (
 
 	"github.com/openshift/library-go/pkg/crypto"
 
+	v3 "github.com/tigera/api/pkg/apis/projectcalico/v3"
 	operator "github.com/tigera/operator/api/v1"
+	"github.com/tigera/operator/pkg/controller/status"
 
 	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
 
 // ExpectResourceCreated asserts that the given object is created,
@@ -215,4 +220,33 @@ func AssertNodesUnchanged(c kubernetes.Interface, nodes ...*v1.Node) error {
 		}
 	}
 	return nil
+}
+
+// DeleteAllowTigeraTierAndExpectWait deletes the tier resource and expects the Reconciler issues a degraded status, waiting for
+// the tier to become available before progressing its status further. Assumes that mockStatus has any required initial status
+// progression expectations set, and that the Reconciler utilizes the mockStatus object. Assumes the tier resource has been created.
+func DeleteAllowTigeraTierAndExpectWait(ctx context.Context, c client.Client, r reconcile.Reconciler, mockStatus *status.MockStatus) {
+	err := c.Delete(ctx, &v3.Tier{ObjectMeta: metav1.ObjectMeta{Name: "allow-tigera"}})
+	Expect(err).ShouldNot(HaveOccurred())
+	mockStatus.On("SetDegraded", operator.ResourceNotReady, "Waiting for allow-tigera tier to be created, see the 'tiers' TigeraStatus for more information", "tiers.projectcalico.org \"allow-tigera\" not found", mock.Anything).Return()
+	_, err = r.Reconcile(ctx, reconcile.Request{})
+	Expect(err).ShouldNot(HaveOccurred())
+	mockStatus.AssertExpectations(GinkgoT())
+}
+
+// ExpectWaitForTierWatch expects the Reconciler issues a degraded status, waiting for a Tier watch to be established.
+// Assumes that mockStatus has any required initial status progression expectations set, and that the Reconciler utilizes
+// the mockStatus object.
+func ExpectWaitForTierWatch(ctx context.Context, r reconcile.Reconciler, mockStatus *status.MockStatus) {
+	ExpectWaitForWatch(ctx, r, mockStatus, "Waiting for Tier watch to be established")
+}
+
+// ExpectWaitForWatch expects the Reconciler issues a degraded status, waiting for a watch to be established.
+// Assumes that mockStatus has any required initial status progression expectations set, and that the Reconciler utilizes
+// the mockStatus object.
+func ExpectWaitForWatch(ctx context.Context, r reconcile.Reconciler, mockStatus *status.MockStatus, message string) {
+	mockStatus.On("SetDegraded", operator.ResourceNotReady, message, mock.Anything, mock.Anything).Return()
+	_, err := r.Reconcile(ctx, reconcile.Request{})
+	Expect(err).ShouldNot(HaveOccurred())
+	mockStatus.AssertExpectations(GinkgoT())
 }
